Use webpack-dev-server
