+++ app/index.html
@@ -12,7 +12,7 @@
             hideBanner: true
         };
     </script>
-    <script src="../node_modules/phaser/build/phaser.min.js"></script>
+    <script src="dist/phaser.js"></script>
     <script src="dist/bundle.js"></script>
     <script>
         (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
