+++ package.json
@@ -1,24 +1,30 @@
 {
-  "private": true,
-  "name": "www.7balls.game",
-  "homepage": "http://www.7balls.game",
-  "scripts": {
-    "clear": "rm -rf node_modules/ app/dist/",
-    "build": "webpack --profile"
-  },
-  "devDependencies": {
-    "babel-core": "^6.9.1",
-    "babel-loader": "^6.2.4",
-    "babel-polyfill": "^6.7.4",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-preset-stage-0": "^6.5.0",
-    "debug": "^2.2.0",
-    "phaser": "^2.4.8",
-    "webpack": "^1.12.0"
-  },
-  "author": {
-    "name": "Piotr Kowalski",
-    "email": "piecioshka@gmail.com",
-    "url": "http://piecioshka.pl/"
-  }
+    "private": true,
+    "name": "www.7balls.game",
+    "homepage": "http://www.7balls.game",
+    "scripts": {
+        "clear": "rm -rf node_modules/ npm-debug.log app/dist/",
+        "build": "webpack --profile",
+        "watch": "webpack -w",
+        "dev": "webpack-dev-server --content-base app"
+    },
+    "devDependencies": {
+        "babel-core": "^6.9.1",
+        "babel-loader": "^6.2.4",
+        "babel-polyfill": "^6.7.4",
+        "babel-preset-es2015": "^6.9.0",
+        "babel-preset-stage-0": "^6.5.0",
+        "debug": "^2.2.0",
+        "expose-loader": "^0.7.1",
+        "imports-loader": "^0.6.5",
+        "phaser": "^2.4.8",
+        "pixi": "^0.3.1",
+        "script-loader": "^0.7.0",
+        "webpack": "^1.12.0"
+    },
+    "author": {
+        "name": "Piotr Kowalski",
+        "email": "piecioshka@gmail.com",
+        "url": "http://piecioshka.pl/"
+    }
 }
